requirements.txt
Flask==2.2.5
flask-cors==3.0.10
google-generativeai==0.5.4
gunicorn
